third_party/WebKit/LayoutTests/platform/mac/fast/events/platform-wheelevent-paging-y-in-scrolling-page-expected.txt

PASS event.wheelDeltaY is window.givenScrollTop * -3
PASS event.wheelDeltaX is 0
PASS event.wheelDelta is window.givenScrollTop * -3
FAIL document.scrollingElement.scrollTop should be 525. Was 560.
PASS document.scrollingElement.scrollLeft is window.expectedScrollLeft
PASS successfullyParsed is true

TEST COMPLETE
